requirements.txt
asgiref==3.3.4
Django==3.2.3
django-cors-headers==3.7.0
django-filter==2.4.0
django-summernote==0.8.11.6
djangorestframework==3.12.4
djangorestframework-simplejwt==4.6.0
Pillow==8.2.0
psycopg2==2.9.1
PyJWT==2.1.0
pytz==2021.1
sqlparse==0.4.1
